Spring ai

+++ book-service/src/main/java/com/guimathis/proxy/ExchangeClient.java
@@ -6,9 +6,11 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
+import java.math.BigDecimal;
+
 @FeignClient(name = "exchange-service", url = "${EXCHANGE_SERVICE_URL:http://localhost:8000}")
 public interface ExchangeClient {
 
     @GetMapping(value = "exchange-service/{value}/{from}/{to}", produces = MediaType.APPLICATION_JSON_VALUE)
-    ExchangeDTO getExchange(@PathVariable Double value, @PathVariable String from, @PathVariable String to);
+    ExchangeDTO getExchange(@PathVariable BigDecimal value, @PathVariable String from, @PathVariable String to);
 }
